Write utility methods for blur image detection
1. compute_directional_kurtosis.py computes the
directional kurtosis of an image and calculates
the score as a feature that might be useful for
detecting blur.

2. compute_local_filter.py is currently a stub
and will be implemented in the future.

3. compute_power_spectrum.py computes the power
spectrum of an image and calculates the score
as a feature. Based on preliminary tests, this
appears to be a good feature.

4. compute_tail_heaviness.py computes the tail
heaviness of an image and calculates the score.

5. naive_bayes_classifier.py is a stub for
a naive bayes classifier that will be used.

6. original_blur_detector.py previous
implementation that uses fft.

Plan to combine the features and feed into
training the naive bayes classifier.